Adapter in Adaoter Ordner, and some name refactoring
