+++ app/src/main/java/org/secuso/privacyfriendlynetmonitor/Activities/HistoryActivity.java
@@ -54,9 +54,6 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v7.widget.Toolbar;
@@ -64,10 +61,10 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ExpandableListView;
-import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import org.secuso.privacyfriendlynetmonitor.Activities.Adapter.ExpandableHistoryListAdapter;
 import org.secuso.privacyfriendlynetmonitor.ConnectionAnalysis.Collector;
 import org.secuso.privacyfriendlynetmonitor.DatabaseUtil.DBApp;
 import org.secuso.privacyfriendlynetmonitor.DatabaseUtil.DaoSession;
@@ -172,7 +169,7 @@ public class HistoryActivity extends BaseActivity {
                 String appName = tx1.getText().toString();
                 TextView tx2 = (TextView) view.findViewById(R.id.historyGroupSubtitle);
                 String appSubName = tx2.getText().toString();
-                Intent myIntent = new Intent(HistoryActivity.this, AppReport_Detail.class);
+                Intent myIntent = new Intent(HistoryActivity.this, AppConnections_Chart.class);
                 myIntent.putExtra("AppName", appName);
                 myIntent.putExtra("AppSubName", appSubName);
                 startActivity(myIntent);
