+++ app/src/main/res/layout/fragment_charts.xml
